[Feat] 인터페이스 - 구현체 구조 삭제  #353
이유는 구현체를 변경할 계획이 없기 때문에
추가로 상수를 사용해서 변수 관리
registActiveQueue
deleteQueue
checkActiveQueue
추가
